food-context.ts back under its file-size budget: portion quantity moves to its owner
server/handlers/food-context.ts was 1486 lines against a 1450 budget. It is now
1371. The 114 lines that came out are adjustFoodsForSegment and its private
helper normaliseWordNumbers, moved verbatim to server/portion-memory.ts.

The food scanner answers WHICH foods a message names. adjustFoodsForSegment
answers HOW MUCH of each — and that is portion-memory's whole subject. The
evidence is in the function body: of the six things it calls, three already
lived in portion-memory (classifyPortionUnit reads the unit word,
personalPortionFor supplies the learned portion when the client said nothing,
scalePortionDescription writes the result back out). It sat in the handler only
because that is where it was first called from. The handler now imports it like
any other caller, and the two names it no longer needs (personalPortionFor,
classifyPortionUnit, scalePortionDescription, PortionStat, portionDefaultCount)
are gone from its import list.

No new module: the modules counter is at 239/239 with zero headroom, so this had
to go to an existing rightful owner. portion-memory.ts is 271 -> 397 lines
against the 1200 default; food-scanner.ts is untouched. No import cycle —
food-scanner reaches portion-memory only through a dynamic import inside a
function, so the new static edge runs one way.

Behaviour preserved and proved, not asserted:
  - gap-tests 335/335, which owns the portion-unit battery (SA end-to-end cases,
    ai-origin tagging, vague quantities, size multipliers) — this is the suite
    that would break if the move changed an answer;
  - production-parity 142/142, food-scanner-tests 48/48,
    tracking-contract-tests green, tsc clean.

One test changed, and it is stronger than what it replaces. The invariant
"memory only fills silence" was checked by running a regex over
food-context.ts source — an assertion that passed on the presence of a line
rather than on what the line did, and which would have kept passing had the
condition been inverted. adjustFoodsForSegment now lives in the same module
that test block already imports, so the rule is asserted where it is decided:
three ways a client can state an amount (explicit "3 eggs", vague "some eggs",
size "big plate of eggs"), each of which must beat their 280 kcal history,
against the one silent case where history is allowed to win. Negative control:
deleting the !explicitQty && !vagueQty && sizeMultiplier === 1 gate fails the
test on the first speech case; restored, 1053/1053.

Governors, main@1c25ce3 -> here:
  file-size    food-context.ts 1486 -> 1371 (budget 1450), off the failing list.
               media.ts 1572, routes.ts 1500, utils.ts 1386 unchanged and
               pre-existing. No file newly over budget.
  architecture messageDeciders 32, looksLikePredicates 21, authorshipPoints 424
               — identical. modules 239, no rise. ownership OK.
  names 97/97, reach 8 paths, SAST 64/64 — all unchanged.

No budget raised, nothing suppressed, no customer copy or coaching policy
touched. Codex's #140 surface (routes.ts, steps.ts, workout.ts, safety.ts) is
not in this diff.
